+++ src/xmlgen.h
@@ -53,7 +53,7 @@ class XMLCodeGenerator
     QCString m_external;
     int m_lineNumber;
     bool m_isMemberRef;
-    int m_col;
+    size_t m_col;
 
     bool m_insideCodeLine;
     bool m_normalHLNeedStartTag;
